requirements.txt
python-dotenv~=1.1.0
utils~=1.0.2
pytest~=8.3.5
requests~=2.32.3
Faker~=37.1.0
playwright~=1.52.0
allure-python-commons~=2.14.3
pydantic~=2.11.7